Respond to review comments
- Use size_t to avoid static_casts in the C-methods
- Uses the typedefs for radex_client_t and
  radex_[incoming/outgoing]_handle_t throughout the code

+++ include/radex/macros/client_c_declaration_macros.h
@@ -21,7 +21,7 @@
 #define RADEX_DECLARE_C_CLIENT_WAIT_FOR_SCALAR(CTYPE, TYPE_TAG) \
   int radex_client_wait_for_##TYPE_TAG(radex_client_t client, \
                                        radex_incoming_handle_t* handle, \
-                                       CTYPE* out_value, int timeout_ms);
+                                       CTYPE* out_value, size_t timeout_ms);
 
 // Tensor methods declarations
 
@@ -29,21 +29,21 @@
 #define RADEX_DECLARE_C_CLIENT_PUT_TENSOR(CTYPE, TYPE_TAG) \
   int radex_client_put_tensor_##TYPE_TAG(radex_client_t client, \
                                          radex_outgoing_handle_t* handle, \
-                                         const CTYPE* data, int rank, const int* dims);
+                                         const CTYPE* data, size_t rank, const size_t* dims);
 
 /// Declare a get_tensor function for a given C type
 #define RADEX_DECLARE_C_CLIENT_GET_TENSOR(CTYPE, TYPE_TAG) \
   int radex_client_get_tensor_##TYPE_TAG(radex_client_t client, \
                                          radex_incoming_handle_t* handle, \
-                                         CTYPE* out_data, int max_elements, \
-                                         int* out_rank, int* out_dims, int max_dims);
+                                         CTYPE* out_data, size_t max_elements, \
+                                         size_t* out_rank, size_t* out_dims, size_t max_dims);
 
 /// Declare a wait_for_tensor function for a given C type
 #define RADEX_DECLARE_C_CLIENT_WAIT_FOR_TENSOR(CTYPE, TYPE_TAG) \
   int radex_client_wait_for_tensor_##TYPE_TAG(radex_client_t client, \
                                               radex_incoming_handle_t* handle, \
-                                              CTYPE* out_data, int max_elements, \
-                                              int* out_rank, int* out_dims, int max_dims, \
-                                              int timeout_ms);
+                                              CTYPE* out_data, size_t max_elements, \
+                                              size_t* out_rank, size_t* out_dims, size_t max_dims, \
+                                              size_t timeout_ms);
 
 #endif // RADEX_CLIENT_C_DECLARATION_MACROS_H
